requirements.txt
defusedxml==0.4.1
Django==1.10.2
django-allauth==0.27.0
django-apptemplates==1.2
django-bootstrap3==7.1.0
django-debug-toolbar==1.6
django-extensions==1.7.4
django-paypal==0.3.3
oauthlib==2.0.0
python3-openid==3.0.10
pytz==2016.7
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
sqlparse==0.2.1
